requirements.txt
astroid==3.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.20.1
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.0
httplib2==0.22.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pylint==3.1.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
